# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para confeitaria premium Docinho & Cia
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SiteSettings {
  id: number;
  shopName: string;
  /** @nullable */
  logoUrl: string | null;
  primaryColor: string;
  secondaryColor: string;
  accentColor: string;
  whatsappNumber: string;
  whatsappMessage: string;
  /** @nullable */
  address: string | null;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  instagram: string | null;
  /** @nullable */
  facebook: string | null;
  /** @nullable */
  tiktok: string | null;
  /** @nullable */
  footerText?: string | null;
}

export interface SiteSettingsInput {
  shopName?: string;
  /** @nullable */
  logoUrl?: string | null;
  primaryColor?: string;
  secondaryColor?: string;
  accentColor?: string;
  whatsappNumber?: string;
  whatsappMessage?: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  facebook?: string | null;
  /** @nullable */
  tiktok?: string | null;
  /** @nullable */
  footerText?: string | null;
}

/**
 * JSON blob with section-specific content
 */
export type SectionContent = { [key: string]: unknown };

export interface Section {
  id: number;
  slug: string;
  title: string;
  visible: boolean;
  sortOrder: number;
  /** JSON blob with section-specific content */
  content: SectionContent;
}

export type SectionInputContent = { [key: string]: unknown };

export interface SectionInput {
  title: string;
  slug: string;
  visible?: boolean;
  sortOrder?: number;
  content?: SectionInputContent;
}

export type SectionUpdateContent = { [key: string]: unknown };

export interface SectionUpdate {
  title?: string;
  visible?: boolean;
  sortOrder?: number;
  content?: SectionUpdateContent;
}

export interface ReorderInput {
  ids: number[];
}

export interface Sweet {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  imageUrl: string | null;
  sortOrder: number;
  visible: boolean;
}

export interface SweetInput {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder?: number;
  visible?: boolean;
}

export interface SweetUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder?: number;
  visible?: boolean;
}

export interface Cake {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl: string | null;
  sortOrder: number;
  visible: boolean;
}

export interface CakeInput {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder?: number;
  visible?: boolean;
}

export interface CakeUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder?: number;
  visible?: boolean;
}

export interface Kit {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  imageUrl: string | null;
  sortOrder: number;
  visible: boolean;
}

export interface KitInput {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder?: number;
  visible?: boolean;
}

export interface KitUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder?: number;
  visible?: boolean;
}

export interface Highlight {
  id: number;
  imageUrl: string;
  /** @nullable */
  caption?: string | null;
  sortOrder: number;
  visible: boolean;
}

export interface HighlightInput {
  imageUrl: string;
  /** @nullable */
  caption?: string | null;
  sortOrder?: number;
  visible?: boolean;
}

export interface HighlightUpdate {
  imageUrl?: string;
  /** @nullable */
  caption?: string | null;
  sortOrder?: number;
  visible?: boolean;
}

export interface Testimonial {
  id: number;
  name: string;
  text: string;
  /** @nullable */
  photoUrl?: string | null;
  sortOrder: number;
  visible: boolean;
}

export interface TestimonialInput {
  name: string;
  text: string;
  /** @nullable */
  photoUrl?: string | null;
  sortOrder?: number;
  visible?: boolean;
}

export interface TestimonialUpdate {
  name?: string;
  text?: string;
  /** @nullable */
  photoUrl?: string | null;
  sortOrder?: number;
  visible?: boolean;
}

export interface AdminCredentials {
  login: string;
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export interface ImageUploadInput {
  /** Base64 encoded image data */
  data: string;
  filename: string;
}

export interface ImageUploadResult {
  url: string;
}