# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API para confeitaria premium Docinho & Cia
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sections:
    get:
      operationId: getSections
      tags: [sections]
      summary: Get all sections ordered
      responses:
        "200":
          description: List of sections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Section"
    post:
      operationId: createSection
      tags: [sections]
      summary: Create a custom section
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionInput"
      responses:
        "201":
          description: Created section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
  /sections/reorder:
    put:
      operationId: reorderSections
      tags: [sections]
      summary: Reorder sections via drag and drop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          description: Sections reordered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Section"
  /sections/{id}:
    put:
      operationId: updateSection
      tags: [sections]
      summary: Update a section content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionUpdate"
      responses:
        "200":
          description: Updated section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
    delete:
      operationId: deleteSection
      tags: [sections]
      summary: Delete a custom section
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /cakes:
    get:
      operationId: getCakes
      tags: [cakes]
      summary: List all cakes
      responses:
        "200":
          description: List of cakes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cake"
    post:
      operationId: createCake
      tags: [cakes]
      summary: Add a new cake
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CakeInput"
      responses:
        "201":
          description: Created cake
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cake"
  /cakes/{id}:
    put:
      operationId: updateCake
      tags: [cakes]
      summary: Update a cake
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CakeUpdate"
      responses:
        "200":
          description: Updated cake
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cake"
    delete:
      operationId: deleteCake
      tags: [cakes]
      summary: Delete a cake
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /kits:
    get:
      operationId: getKits
      tags: [kits]
      summary: List all kits
      responses:
        "200":
          description: List of kits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Kit"
    post:
      operationId: createKit
      tags: [kits]
      summary: Add a new kit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KitInput"
      responses:
        "201":
          description: Created kit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kit"
  /kits/{id}:
    put:
      operationId: updateKit
      tags: [kits]
      summary: Update a kit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KitUpdate"
      responses:
        "200":
          description: Updated kit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Kit"
    delete:
      operationId: deleteKit
      tags: [kits]
      summary: Delete a kit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /highlights:
    get:
      operationId: getHighlights
      tags: [highlights]
      summary: List all highlight slides
      responses:
        "200":
          description: List of highlights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Highlight"
    post:
      operationId: createHighlight
      tags: [highlights]
      summary: Add a highlight slide
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HighlightInput"
      responses:
        "201":
          description: Created highlight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Highlight"
  /highlights/{id}:
    put:
      operationId: updateHighlight
      tags: [highlights]
      summary: Update a highlight
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HighlightUpdate"
      responses:
        "200":
          description: Updated highlight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Highlight"
    delete:
      operationId: deleteHighlight
      tags: [highlights]
      summary: Delete a highlight
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
...
  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Check admin session
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated
...
components:
  schemas:
...
    Section:
      type: object
      required: [id, slug, title, visible, sortOrder, content]
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        visible:
          type: boolean
        sortOrder:
          type: integer
        content:
          type: object
          description: JSON blob with section-specific content
    SectionInput:
      type: object
      required: [title, slug]
      properties:
        title:
          type: string
        slug:
          type: string
        visible:
          type: boolean
        sortOrder:
          type: integer
        content:
          type: object
    SectionUpdate:
      type: object
      properties:
        title:
          type: string
        visible:
          type: boolean
        sortOrder:
          type: integer
        content:
          type: object
    ReorderInput:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: integer
...
    Cake:
      type: object
      required: [id, name, imageUrl, sortOrder, visible]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        visible:
          type: boolean
    CakeInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        visible:
          type: boolean
    CakeUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        visible:
          type: boolean
    Kit:
      type: object
      required: [id, name, imageUrl, sortOrder, visible]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        visible:
          type: boolean
    KitInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        visible:
          type: boolean
    KitUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        visible:
          type: boolean
    Highlight:
      type: object
      required: [id, imageUrl, sortOrder, visible]
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        caption:
          type: ["string", "null"]
        sortOrder:
          type: integer
        visible:
          type: boolean
    HighlightInput:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
        caption:
          type: ["string", "null"]
        sortOrder:
          type: integer
        visible:
          type: boolean
    HighlightUpdate:
      type: object
      properties:
        imageUrl:
          type: string
        caption:
          type: ["string", "null"]
        sortOrder:
          type: integer
        visible:
          type: boolean
...
    AdminCredentials:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
        password:
          type: string
    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean
...