# lib/api-zod/src/generated/types/highlightUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para confeitaria premium Docinho & Cia
 * OpenAPI spec version: 0.1.0
 */

export interface HighlightUpdate {
  imageUrl?: string;
  /** @nullable */
  caption?: string | null;
  sortOrder?: number;
  visible?: boolean;
}