# lib/api-zod/src/generated/types/kitUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API para confeitaria premium Docinho & Cia
 * OpenAPI spec version: 0.1.0
 */

export interface KitUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder?: number;
  visible?: boolean;
}